Resolves #12 + Minor code factoring
