+++ Concentration/Concentration/Card.swift
@@ -9,18 +9,32 @@
 import Foundation
 
 struct Card {
+    
+    // MARK: - Public types
+    
     var isFaceUp = false
+    
     var isMatched = false
+    
+    // MARK: - Semipublic types
+    
     private(set) var identifier: Int
     
+    // MARK: - Private types
+    
     private static var identifierGlobalCounter = 0
+    
+    // MARK: - Init
+    
+    init() {
+        self.identifier = Card.getUniqueIdentifier()
+    }
+    
+    // MARK: - Private methods
     
     private static func getUniqueIdentifier() -> Int {
         identifierGlobalCounter += 1
         return identifierGlobalCounter
     }
     
-    init() {
-        self.identifier = Card.getUniqueIdentifier()
-    }
 }
