+++ Concentration/Concentration/ViewController/ViewController.swift
@@ -10,9 +10,12 @@ import UIKit
 
 class ViewController: UIViewController {
     
+    // MARK: - Private types
+    
     private lazy var game = Concentration(numberOfPairsOfCards: numberOfPairsOfCards)
     
     private var theme: Theme = Theme.randomTheme() {
+        
         didSet {
             emojiSet = theme.emojiSet
             self.view.backgroundColor = theme.backgroundColor
@@ -24,7 +27,9 @@ class ViewController: UIViewController {
             for button in cardButtons {
                 button.backgroundColor = theme.cardColor
             }
+            
         }
+        
     }
     
     private lazy var emojiSet: [String] = theme.emojiSet
@@ -36,36 +41,57 @@ class ViewController: UIViewController {
     }
     
     @IBOutlet private var cardButtons: [UIButton]!
+    
     @IBOutlet private weak var flipCountLabel: UILabel!
+    
     @IBOutlet private weak var scoreLabel: UILabel!
+    
     @IBOutlet private weak var newGameButton: UIButton!
     
+    // MARK: - Semiprivate types
+    
     private(set) var flipCount = 0 {
+        
         didSet {
             flipCountLabel.text = String.StaticTexts.flipsLabelText + "\(flipCount)"
         }
+        
     }
     
     private(set) var gameScore = 0 {
+        
         didSet {
             scoreLabel.text = String.StaticTexts.scoreLabelText + "\(gameScore)"
         }
+        
     }
-
-    @IBAction private func touch(_ sender: UIButton) {
-        if let cardNumber = cardButtons.index(of: sender) {
-            game.chooseCard(at: cardNumber)
-            updateViewFromModel()
-        }
-        else {
-            print("Chosen card was not in cardButtons")
-        }
-    }
+    
+    // MARK: - Public methods
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
         startNewGame()
+    }
+    
+    private func emoji(for card: Card) -> String {
+        
+        if emoji[card.identifier] == nil, emojiSet.count > 0 {
+            let randomIndex = Int.random(in: 0..<emojiSet.count)
+            emoji[card.identifier] = emojiSet.remove(at: randomIndex)
+        }
+        return emoji[card.identifier] ?? "?"
+    }
+    
+    // MARK: - Private methods
+
+    @IBAction private func touch(_ sender: UIButton) {
+        
+        if let cardNumber = cardButtons.index(of: sender) {
+            game.chooseCard(at: cardNumber)
+            updateViewFromModel()
+        } else {
+            print("Chosen card was not in cardButtons")
+        }
         
     }
     
@@ -93,28 +119,5 @@ class ViewController: UIViewController {
         flipCount = game.flips
         gameScore = game.score
     }
-    
-    private func emoji(for card: Card) -> String {
-        
-        if emoji[card.identifier] == nil, emojiSet.count > 0 {
-            let randomIndex = Int.random(in: 0..<emojiSet.count)
-            emoji[card.identifier] = emojiSet.remove(at: randomIndex)
-        }
-        return emoji[card.identifier] ?? "?"
-    }
-}
 
-extension Int {
-    
-    var arc4random: Int {
-        if self > 0 {
-            return Int.random(in: 0..<self)
-        } else if self < 0 {
-            return -Int.random(in: 0..<abs(self))
-        } else {
-            return 0
-        }
-        
-    }
-    
 }
